+++ src/components/chrome/ChromeFields.js
@@ -30,6 +30,10 @@ export class ChromeFields extends ReactCSS.Component {
           paddingLeft: '6px',
           width: '100%',
         },
+        alpha: {
+          paddingLeft: '6px',
+          width: '100%',
+        },
         toggle: {
           width: '32px',
           textAlign: 'right',
@@ -39,7 +43,7 @@ export class ChromeFields extends ReactCSS.Component {
           marginRight: '-4px',
           marginTop: '12px',
           cursor: 'pointer',
-          position: 'relative'
+          position: 'relative',
         },
         iconHighlight: {
           position: 'absolute',
@@ -73,6 +77,11 @@ export class ChromeFields extends ReactCSS.Component {
               marginTop: '12px',
             },
           },
+        },
+      },
+      'disableAlpha': {
+        alpha: {
+          display: 'none',
         },
       },
     }
@@ -148,12 +157,12 @@ export class ChromeFields extends ReactCSS.Component {
     }
   }
 
-  showHighlight = () => {
-    this.refs.iconHighlight.style.display = 'block'
+  showHighlight = (e) => {
+    e.target.style.background = "#eee"
   }
 
-  hideHighlight = () => {
-    this.refs.iconHighlight.style.display = 'none'
+  hideHighlight = (e) => {
+    e.target.style.background = "transparent"
   }
 
   render(): any {
@@ -161,7 +170,7 @@ export class ChromeFields extends ReactCSS.Component {
     if (this.state.view === 'hex') {
       fields = <div is="fields" className="flexbox-fix">
         <div is="field">
-          <EditableInput is="Input" label="hex" value={ '#' + this.props.hex } onChange={ this.handleChange }/>
+          <EditableInput is="Input" label="hex" value={ this.props.hex } onChange={ this.handleChange }/>
         </div>
       </div>
     } else if (this.state.view === 'rgb') {
@@ -175,7 +184,7 @@ export class ChromeFields extends ReactCSS.Component {
         <div is="field">
           <EditableInput is="Input" label="b" value={ this.props.rgb.b } onChange={ this.handleChange } />
         </div>
-        <div is="field">
+        <div is="alpha">
           <EditableInput is="Input" label="a" value={ this.props.rgb.a } arrowOffset={ .01 } onChange={ this.handleChange } />
         </div>
       </div>
@@ -190,7 +199,7 @@ export class ChromeFields extends ReactCSS.Component {
         <div is="field">
           <EditableInput is="Input" label="l" value={ Math.round(this.props.hsl.l * 100) + '%' } onChange={ this.handleChange } />
         </div>
-        <div is="field">
+        <div is="alpha">
           <EditableInput is="Input" label="a" value={ this.props.hsl.a } arrowOffset={ .01 } onChange={ this.handleChange } />
         </div>
       </div>
@@ -201,11 +210,15 @@ export class ChromeFields extends ReactCSS.Component {
         { fields }
         <div is="toggle">
           <div is="icon" onClick={ this.toggleViews } ref="icon">
-            <svg style={{ width:'24px', height:'24px', }} viewBox="0 0 24 24" onMouseOver={ this.showHighlight } onMouseEnter={ this.showHighlight } onMouseOut={ this.hideHighlight }>
-              <path fill="#333" d="M12,18.17L8.83,15L7.42,16.41L12,21L16.59,16.41L15.17,15M12,5.83L15.17,9L16.58,7.59L12,3L7.41,7.59L8.83,9L12,5.83Z" />
+            <svg style={{ width:'24px', height:'24px', border: '1px transparent solid', borderRadius: '5px' }}
+              viewBox="0 0 24 24"
+              onMouseOver={ this.showHighlight }
+              onMouseEnter={ this.showHighlight }
+              onMouseOut={ this.hideHighlight }>
+              <path ref="iconUp" fill="#333" d="M12,5.83L15.17,9L16.58,7.59L12,3L7.41,7.59L8.83,9L12,5.83Z" />
+              <path ref="iconDown" fill="#333" d="M12,18.17L8.83,15L7.42,16.41L12,21L16.59,16.41L15.17,15Z"/>
             </svg>
           </div>
-          <div is="iconHighlight" ref="iconHighlight" />
         </div>
       </div>
     )